Conflicts:
	src/main/java/com/mii/cvonlinerestserver/controller/LowonganController.java

+++ src/main/java/com/mii/cvonlinerestserver/dao/LowonganDAO.java
@@ -24,7 +24,7 @@ public class LowonganDAO {
 		return lowonganRepository.findAll();
 	}
 	
-	public Lowongan findOne(Long id) {
+	public Lowongan findOne(Integer id) {
 		return lowonganRepository.findOne(id);
 	}
